# src/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8082",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
        "/api/recipes/{id}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Obtener receta",
                "description": "Obtiene una receta por su ID",
                "operationId": "getRecipe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la receta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Receta no encontrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Recipe"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Receta encontrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Recipe"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Actualizar receta",
                "description": "Actualiza una receta existente por ID",
                "operationId": "updateRecipe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la receta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Recipe"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Receta actualizada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Recipe"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Receta no encontrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Recipe"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/recipes": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Listar recetas activas",
                "description": "Obtiene todas las recetas activas",
                "operationId": "getAllRecipes",
                "responses": {
                    "200": {
                        "description": "Lista obtenida",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Recipe"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Crear receta",
                "description": "Crea una nueva receta de yogurt",
                "operationId": "createRecipe",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Recipe"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Receta creada correctamente",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Recipe"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Recipe"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/recipes/{id}/deactivate": {
            "patch": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Desactivar receta",
                "description": "Desactiva una receta (soft delete)",
                "operationId": "deactivateRecipe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la receta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receta desactivada"
                    }
                }
            }
        },
        "/api/recipes/{id}/activate": {
            "patch": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Activar receta",
                "description": "Activa una receta previamente desactivada",
                "operationId": "activateRecipe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la receta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receta activada"
                    }
                }
            }
        },
        "/api/recipes/search": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Buscar recetas",
                "description": "Busca recetas por palabra clave",
                "operationId": "searchRecipe",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Palabra clave de búsqueda",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "natural"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados encontrados",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Recipe"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parámetro inválido",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Recipe"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Ingredient": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Nombre del ingrediente",
                        "example": "Leche entera"
                    },
                    "quantity": {
                        "type": "number",
                        "description": "Cantidad del ingrediente",
                        "format": "double",
                        "example": 1.0
                    },
                    "unit": {
                        "type": "string",
                        "description": "Unidad de medida",
                        "example": "litros"
                    },
                    "notes": {
                        "type": "string",
                        "description": "Notas adicionales sobre el ingrediente",
                        "example": "Preferiblemente fresca"
                    },
                    "options": {
                        "type": "boolean",
                        "description": "Indica si el ingrediente es opcional",
                        "example": false
                    }
                },
                "description": "Ingrediente utilizado en una receta de yogurt"
            },
            "Recipe": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Nombre de la receta",
                        "example": "Yogurt natural clásico"
                    },
                    "description": {
                        "type": "string",
                        "description": "Descripción de la receta",
                        "example": "Receta tradicional de yogurt casero"
                    },
                    "defaulMilkVolume": {
                        "type": "number",
                        "description": "Volumen de leche en litros",
                        "format": "double",
                        "example": 1.0
                    },
                    "defaulStarterAmount": {
                        "type": "number",
                        "description": "Cantidad de cultivo iniciador (starter) en gramos",
                        "format": "double",
                        "example": 50.0
                    },
                    "heatingTemperature": {
                        "type": "number",
                        "description": "Temperatura de calentamiento en °C",
                        "format": "double",
                        "example": 85.0
                    },
                    "heatingDuration": {
                        "type": "integer",
                        "description": "Duración del calentamiento en minutos",
                        "format": "int32",
                        "example": 30
                    },
                    "innoculationTemperature": {
                        "type": "number",
                        "description": "Temperatura de inoculación en °C",
                        "format": "double",
                        "example": 43.0
                    },
                    "incubationTemperature": {
                        "type": "number",
                        "description": "Temperatura de incubación en °C",
                        "format": "double",
                        "example": 42.0
                    },
                    "minIncubationTime": {
                        "type": "integer",
                        "description": "Tiempo mínimo de incubación en minutos",
                        "format": "int32",
                        "example": 240
                    },
                    "maxIncubationTime": {
                        "type": "integer",
                        "description": "Tiempo máximo de incubación en minutos",
                        "format": "int32",
                        "example": 480
                    },
                    "refrigerationTime": {
                        "type": "integer",
                        "description": "Tiempo de refrigeración en minutos",
                        "format": "int32",
                        "example": 180
                    },
                    "difficulty": {
                        "type": "string",
                        "description": "Nivel de dificultad de la receta",
                        "example": "BEGINNER",
                        "enum": [
                            "BEGINNER",
                            "INTERMEDIATE",
                            "ADVANCED"
                        ]
                    },
                    "tips": {
                        "type": "string",
                        "description": "Consejos adicionales",
                        "example": "Mantener temperatura constante durante la incubación"
                    },
                    "ingredients": {
                        "type": "array",
                        "description": "Lista de ingredientes de la receta",
                        "items": {
                            "$ref": "#/components/schemas/Ingredient"
                        }
                    }
                },
                "description": "DTO para la creación y actualización de recetas de yogurt"
            },
...
        }
    }
}